New regex to match link on new website
Shure updated their website and this recipe was failing due to that.  Good news, the file is consistent now.  Bad news, the version number is harder to find.  I updated the regex to match the link and version on the new website/file structure they have setup.

This is tested and working for me.