Add installcheck_script, postinstall_script, & RECEIPT_PATH var
pkg receipts don't have any versions, so even though update_for is tied only to specific version, if we don't have a file to reference for this from items installed (which I haven't been able to find), we need to make our own.

Write a receipt file with the version installed contained inside as part of postinstall.  Refer to this file in installcheck_script to determine if installed version (if present) is not the same, and if true also make sure the installed version is less than what munki is checking.